Stabilize answer submission ownership and scoring state
Constraint: CodeRabbit review required actual question IDs in the public API and no duplicate LLM scoring re-entry while answers are pending. Existing Flyway schema uses ddl-auto validate, so processing was modeled with existing status storage instead of adding lease columns.
Rejected: Keep treating questionId as turn | It leaked an internal sequencing value through the API and made tests submit synthetic IDs.
Confidence: high
Scope-risk: moderate
Directive: Keep judgment questionId values as internal turns unless the judgment raw model contract is explicitly changed; map to public message IDs at the interview API boundary.
Tested: compileTestJava; targeted interview controller/service tests; full gradlew test; git diff --check
Not-tested: Cross-instance lease expiry beyond immediate status release on failure

+++ src/main/java/com/careerdungeon/domain/interview/dto/InterviewAnswerEvaluationResponse.java
@@ -1,7 +1,7 @@
 package com.careerdungeon.domain.interview.dto;
 
 public record InterviewAnswerEvaluationResponse(
-        int questionId,
+        Long questionId,
         int score,
         String feedback
 ) {
